contract/contracts/FundSphere/src/test.rs
// Import necessary modules and traits for testing
use soroban_sdk::{Context, Env, ReplyError};
use your_contract_module::{create_campaign, fund_to_campaign, get_total_amount_collected, get_funders, get_campaigns, Campaign}; // Import your contract module and functions

// Mock Context and Env for testing
struct MockContext;

impl Context for MockContext {}

impl Env for MockEnv {
    // Implement necessary functions from the Env trait for testing
}

// Test create_campaign function
#[test]
fn test_create_campaign() {
    let ctx = MockContext;
    let title = "Test Campaign";
    let description = "This is a test campaign";
    let target = 1000;
    let deadline = 1735689600; // Unix timestamp for January 1, 2025
    let image = "test_image.jpg";

    let result = create_campaign(&ctx, "", title, description, target, deadline, image);
    assert!(result.is_ok());
}

// Test fund_to_campaign function
#[test]
fn test_fund_to_campaign() {
    let ctx = MockContext;
    let campaign_id = 1;

    let result = fund_to_campaign(&ctx, campaign_id);
    assert!(result.is_ok());
}

// Test get_total_amount_collected function
#[test]
fn test_get_total_amount_collected() {
    let ctx = MockContext;
    let campaign_id = 1;

    let result = get_total_amount_collected(&ctx, campaign_id);
    assert!(result.is_ok());
}

// Test get_funders function
#[test]
fn test_get_funders() {
    let ctx = MockContext;
    let campaign_id = 1;

    let result = get_funders(&ctx, campaign_id);
    assert!(result.is_ok());
}

// Test get_campaigns function
#[test]
fn test_get_campaigns() {
    let ctx = MockContext;

    let result = get_campaigns(&ctx);
    assert!(result.is_ok());
}
